M2 on STM32: AD7606 reading correctly, and three silent-failure bugs
The AD7606 reads both known signals correctly on the first attempt: V1
swinging -1.2433 to +1.2216 against the +/-1.22 V scope output, and V8
alternating 0.0107 to 2.0087 against the 0-2 V Hantek square. So
SPI_MODE2 is correct on the STM32 as well as the P4 - it had been marked
an unverified placeholder, and two known signals settled it without
needing a sweep.

M0, M1 and M2 are now verified on both targets.

Three bugs found, all of which failed silently:

program firmware.bin wrote nothing and reported success. A raw .bin
carries no address information, so without an explicit 0x08000000
OpenOCD does nothing while printing "Programming Finished". The board
reset cleanly and kept running the previous image, with a correct build
sitting unused on disk. Now programming the ELF, which carries its own
load addresses.

Dropping verify is what hid it. The earlier CRC verify failures looked
like a nuisance to work around, and were actually the symptom of nothing
being written - turning verify off removed the only signal that anything
was wrong. It is back on and stays on.

And newlib-nano's printf omits float support, so %f prints nothing at
all rather than erroring. Every voltage and size in the output came out
blank, which reads as a data problem rather than a linker one. Fixed
with -Wl,-u,_printf_float.

+++ firmware/ecu/include/Ad7606c.h
@@ -47,15 +47,18 @@ enum class Oversampling : uint8_t {
 };
 
 struct Pins {
-    int8_t sck;
-    int8_t miso;      // AD7606C DOUTA
-    int8_t cs;        // software-driven, not hardware SPI CS
-    int8_t convst;    // also drives WR on the usual breakouts
-    int8_t busy;
-    int8_t reset;
-    int8_t range;
-    int8_t os0, os1, os2;
-    int8_t frstdata = -1;  // optional; -1 if not wired
+    // int16_t, not int8_t: STM32duino's pin constants run past 127 (PB10 is
+    // 199), while -1 still means "not wired". The ESP32's numbering happened
+    // to fit in a signed byte; that was luck rather than design.
+    int16_t sck;
+    int16_t miso;      // AD7606C DOUTA
+    int16_t cs;        // software-driven, not hardware SPI CS
+    int16_t convst;    // also drives WR on the usual breakouts
+    int16_t busy;
+    int16_t reset;
+    int16_t range;
+    int16_t os0, os1, os2;
+    int16_t frstdata = -1;  // optional; -1 if not wired
 };
 
 // spiHz defaults to 1 MHz — deliberately slow for bring-up over flying leads.
